@@ -1,20 +1,35 @@
 require("./modules/bootstrap");
 require("./modules/websockets");
-const components = require("./modules/components")("app");
 
+import { createApp } from "vue";
+import mixins from "./mixins";
 import store from "./store";
 import { mapGetters, mapActions } from "vuex";
 
-const app = new Vue({
-    components: { components },
-    el: "#app",
-    store,
-    data: function() {
+import DetailsDocument from "./components/Details/DetailsDocument.vue";
+import DetailsDocuments from "./components/Details/DetailsDocuments.vue";
+import DetailsFeedItem from "./components/Details/DetailsFeedItem.vue";
+import DetailsFolder from "./components/Details/DetailsFolder.vue";
+import DocumentsList from "./components/DocumentsList.vue";
+import FeedItemsList from "./components/FeedItemsList.vue";
+import FoldersTree from "./components/FoldersTree.vue";
+
+createApp({
+    components: {
+        DetailsDocument,
+        DetailsDocuments,
+        DetailsFeedItem,
+        DetailsFolder,
+        DocumentsList,
+        FeedItemsList,
+        FoldersTree,
+    },
+    data: function () {
         return {
-            detailsViewComponent: null
+            detailsViewComponent: null,
         };
     },
-    mounted: function() {
+    mounted: function () {
         const self = this;
 
         self.listenToBroadcast();
@@ -27,16 +42,18 @@ const app = new Vue({
             feedItems: "feedItems/feedItems",
             selectedFeedItems: "feedItems/selectedFeedItems",
             getUnreadItemsFolder: "folders/getUnreadItemsFolder",
-            selectedGroup: "groups/selectedGroup"
-        })
+            selectedGroup: "groups/selectedGroup",
+        }),
     },
     watch: {
-        selectedFolder: function(oldFolder, newFolder) {
+        selectedFolder: function (oldFolder, newFolder) {
             const self = this;
 
-            self.detailsViewComponent = "details-folder";
+            if (newFolder) {
+                self.detailsViewComponent = "details-folder";
+            }
         },
-        selectedDocuments: function(documents) {
+        selectedDocuments: function (documents) {
             const self = this;
 
             if (documents && documents.length > 0) {
@@ -49,7 +66,7 @@ const app = new Vue({
                 self.detailsViewComponent = "details-folder";
             }
         },
-        selectedFeedItems: function(feedItems) {
+        selectedFeedItems: function (feedItems) {
             const self = this;
 
             if (feedItems && feedItems.length > 0) {
@@ -69,7 +86,7 @@ const app = new Vue({
                     self.detailsViewComponent = "details-folder";
                 }
             }
-        }
+        },
     },
     methods: {
         ...mapActions({
@@ -86,32 +103,32 @@ const app = new Vue({
             showGroup: "groups/show",
             updateGroup: "groups/updateProperties",
             resetUnreadCount: "folders/resetUnreadCount",
-            updateUnreadFeedItemsCount: "feedItems/updateUnreadFeedItemsCount"
+            updateUnreadFeedItemsCount: "feedItems/updateUnreadFeedItemsCount",
         }),
 
         /**
          * Listen to broadcast events
          */
-        listenToBroadcast: function() {
+        listenToBroadcast: function () {
             const self = this;
             const userId = document
                 .querySelector('meta[name="user-id"]')
                 .getAttribute("content");
 
             window.Echo.private("App.Models.User." + userId).notification(
-                notification => {
+                (notification) => {
                     switch (notification.type) {
                         case "App\\Notifications\\UnreadItemsChanged":
                             self.updateUnreadFeedItemsCount(notification);
-
-                            if (self.selectedFolder.type === "unread_items") {
-                                self.showFolder();
-                            }
+                            self.showFolder({
+                                deselectDocuments: false,
+                                updateFeedItems: true,
+                            });
                             break;
                         case "App\\Notifications\\DocumentUpdated":
                             self.updateDocument({
                                 documentId: notification.document.id,
-                                newProperties: notification.document
+                                newProperties: notification.document,
                             });
                             break;
                     }
@@ -123,16 +140,23 @@ const app = new Vue({
          * User-action - Selected group has changed
          * @param {*} group
          */
-        onSelectedGroupChanged: function(group) {
+        onSelectedGroupChanged: function (group) {
             const self = this;
 
             self.showGroup(group);
         },
 
         /**
+         * Groups have been loaded
+         */
+        onGroupsLoaded: function () {
+            //
+        },
+
+        /**
          * Folders tree has been loaded
          */
-        onFoldersLoaded: function() {
+        onFoldersLoaded: function () {
             //
         },
 
@@ -140,13 +164,17 @@ const app = new Vue({
          * User-action - Selected folder has changed
          * @param {*} folder
          */
-        onSelectedFolderChanged: function(folder, group) {
+        onSelectedFolderChanged: function (folder, group) {
             const self = this;
+
+            if (!folder) {
+                return;
+            }
 
             if (group && group != self.selectedGroup.id) {
                 self.showGroup(group, folder);
             } else {
-                self.showFolder(folder);
+                self.showFolder({ folder });
             }
 
             self.detailsViewComponent = "details-folder";
@@ -155,7 +183,7 @@ const app = new Vue({
         /**
          * User-action - Something has been dropped into a folder
          */
-        onItemDropped: function(folder) {
+        onItemDropped: function (folder) {
             const self = this;
 
             self.dropIntoFolder(folder);
@@ -164,35 +192,37 @@ const app = new Vue({
         /**
          * User-action - Selected documents has changed
          */
-        onSelectedDocumentsChanged: function(documents) {
+        onSelectedDocumentsChanged: function (documents) {
             const self = this;
 
             self.selectDocuments({ documents: documents });
+
+            document.getElementById("feed-items-list").scrollTop = 0;
         },
 
         /**
          * User-action - Refresh documents list after adding one
          */
-        onDocumentAdded: function() {
+        onDocumentAdded: function () {
             //
         },
 
         /**
          * User-action - Refresh documents list after deleting one (or more)
          */
-        onDocumentsDeleted: function({ folder, documents }) {
+        onDocumentsDeleted: function ({ folder, documents }) {
             const self = this;
 
             self.deleteDocuments({
                 documents: documents,
-                folder: folder
+                folder: folder,
             });
         },
 
         /**
          * User-action - Selected feed items has changed
          */
-        onSelectedFeedItemsChanged: function(feedItems) {
+        onSelectedFeedItemsChanged: function (feedItems) {
             const self = this;
 
             self.selectFeedItems(feedItems);
@@ -201,10 +231,13 @@ const app = new Vue({
         /**
          * User-action - Feed items marked as read
          */
-        onFeedItemsRead: function(data) {
+        onFeedItemsRead: function (data) {
             const self = this;
 
             self.markFeedItemsAsRead(data);
-        }
-    }
-});
+        },
+    },
+})
+    .mixin(mixins)
+    .use(store)
+    .mount("#app");
