@@ -4,12 +4,11 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use App\Models\Traits\Document\AnalysesDocument;
-use App\Models\Traits\HasHistory;
 use App\Models\Traits\HasUrl;
 
 class Document extends Model
 {
-    use AnalysesDocument, HasHistory, HasUrl;
+    use AnalysesDocument, HasUrl;
 
     # --------------------------------------------------------------------------
     # ----| Properties |--------------------------------------------------------
@@ -63,17 +62,6 @@ class Document extends Model
      * @var string
      */
     private $storagePath = null;
-
-    /**
-     * Attributes used to display this model in history
-     *
-     * @var array
-     */
-    protected $historyAttributes = [
-        'url',
-        'title',
-        'favicon'
-    ];
 
     # --------------------------------------------------------------------------
     # ----| Attributes |--------------------------------------------------------
