@@ -5,24 +5,24 @@
 </template>
 <script>
 export default {
-    props: ['datetime', 'onlyDate', 'onlyTime'],
+    props: ["datetime", "onlyDate", "onlyTime"],
     computed: {
-        iso: function() {
-            if(!this.datetime) {
+        iso: function () {
+            if (!this.datetime) {
                 return null;
             }
 
             return new Date(this.datetime).toISOString();
         },
-        formatted: function() {
-            if(this.onlyTime) {
+        formatted: function () {
+            if (this.onlyTime) {
                 return new Date(this.datetime).toLocaleTimeString();
-            } else if(this.onlyDate) {
+            } else if (this.onlyDate) {
                 return new Date(this.datetime).toLocaleDateString();
             }
 
             return new Date(this.datetime).toLocaleString();
-        }
-    }
-}
+        },
+    },
+};
 </script>
