@@ -1,84 +1,26 @@
 <template>
-    <div class="group-item alt">
-        <div class="handle" v-if="movable">
+    <div class="list-item">
+        <div class="handle">
             <svg fill="currentColor" width="16" height="16">
                 <use v-bind:xlink:href="icon('move-v')" />
             </svg>
         </div>
-        <div class="flex-grow p-2">
-            <div class="flex justify-between items-center">
-                <div class="title">{{ group.name }}</div>
-                <div>
-                    <span
-                        v-if="group.pending_users_count > 0"
-                        class="badge warning"
-                        v-bind:title="__('Pending users')"
-                        ><svg
-                            fill="currentColor"
-                            width="16"
-                            height="16"
-                            class="inline mr-2"
-                        >
-                            <use v-bind:xlink:href="icon('group')" /></svg
-                        >{{ group.pending_users_count }}</span
-                    >
-                    <span
-                        class="badge success"
-                        v-bind:title="__('Active users')"
-                        ><svg
-                            fill="currentColor"
-                            width="16"
-                            height="16"
-                            class="inline mr-2"
-                        >
-                            <use v-bind:xlink:href="icon('group')" /></svg
-                        >{{ group.active_users_count }}</span
-                    >
-                    <button
-                        class="info inline ml-2"
-                        v-if="group.pivot"
-                        v-on:click="$emit('selected', group)"
-                    >
-                        <svg
-                            fill="currentColor"
-                            width="16"
-                            height="16"
-                            class="inline mr-2"
-                        >
-                            <use v-bind:xlink:href="icon('update')" /></svg
-                        >{{ __("Edit") }}
-                    </button>
-                    <button
-                        class="success inline ml-2"
-                        v-if="!group.pivot"
-                        v-on:click="$emit('join', group)"
-                    >
-                        <svg
-                            fill="currentColor"
-                            width="16"
-                            height="16"
-                            class="inline mr-2"
-                        >
-                            <use v-bind:xlink:href="icon('join')" /></svg
-                        >{{
-                            group.auto_accept_users ? __("Join") : __("Apply")
-                        }}
-                    </button>
+        <div class="flex-grow flex justify-between">
+            <div>
+                <h2 class="text-black dark:text-white">{{ group.name }}</h2>
+                <div v-if="group.creator" class="mb-1">
+                    {{ __("Created by") }} {{ group.creator.name }}
                 </div>
-            </div>
-            <div class="feed-item-meta" v-if="group.creator">
-                {{ __("Created by") }} {{ group.creator.name }}
-            </div>
-            <div class="flex items-center justify-between mb-1">
-                <div>
+                <div class="mb-1">{{ group.description }}</div>
+                <div class="flex items-center space-x-1">
                     <span
                         v-if="group.pivot"
-                        class="badge mr-2 ml-0"
+                        class="badge"
                         v-bind:class="'group-status-' + group.pivot.status"
                         >{{ status }}</span
                     >
                     <span
-                        class="badge default ml-0"
+                        class="badge default"
                         v-bind:class="{
                             'group-invite-only': group.invite_only,
                         }"
@@ -86,7 +28,7 @@
                         >{{ __("Invite only") }}</span
                     >
                     <span
-                        class="badge ml-0"
+                        class="badge"
                         v-bind:class="{
                             success: group.auto_accept_users,
                             warning: !group.auto_accept_users,
@@ -100,7 +42,59 @@
                     >
                 </div>
             </div>
-            <div class="feed-item-meta">{{ group.description }}</div>
+        </div>
+        <div class="badges">
+            <span
+                v-if="group.pending_users_count > 0"
+                class="badge warning"
+                v-bind:title="__('Pending users')"
+                ><svg
+                    fill="currentColor"
+                    width="16"
+                    height="16"
+                    class="inline mr-2"
+                >
+                    <use v-bind:xlink:href="icon('group')" /></svg
+                >{{ group.pending_users_count }}</span
+            >
+            <span class="badge success" v-bind:title="__('Active users')"
+                ><svg
+                    fill="currentColor"
+                    width="16"
+                    height="16"
+                    class="inline mr-2"
+                >
+                    <use v-bind:xlink:href="icon('group')" /></svg
+                >{{ group.active_users_count }}</span
+            >
+            <button
+                class="info ml-2"
+                v-if="group.pivot"
+                v-on:click="$emit('selected', group)"
+            >
+                <svg
+                    fill="currentColor"
+                    width="16"
+                    height="16"
+                    class="inline mr-2"
+                >
+                    <use v-bind:xlink:href="icon('update')" /></svg
+                >{{ __("Edit") }}
+            </button>
+            <button
+                class="success inline ml-2"
+                v-if="!group.pivot"
+                v-on:click="$emit('join', group)"
+            >
+                <svg
+                    fill="currentColor"
+                    width="16"
+                    height="16"
+                    class="inline mr-2"
+                >
+                    <use v-bind:xlink:href="icon('join')" /></svg
+                >{{ group.auto_accept_users ? __("Join") : __("Apply") }}
+            </button>
         </div>
     </div>
 </template>
