@@ -1,20 +1,17 @@
-import Vue from "vue";
-import Vuex from "vuex";
+import { createStore } from "vuex";
 import folders from "./modules/folders";
 import documents from "./modules/documents";
 import feedItems from "./modules/feeditems";
 import groups from "./modules/groups";
 
-Vue.use(Vuex);
-
 const debug = process.env.NODE_ENV !== "production";
 
-export default new Vuex.Store({
+export default createStore({
     modules: {
         folders,
         documents,
         feedItems,
-        groups
+        groups,
     },
-    strict: debug
+    strict: debug,
 });
