@@ -1,3 +1,3 @@
 <a class="inline-document" 
     target="_blank" rel="noopener noreferrer" href="{{ $document['url'] }}"
->@if(!empty($document['favicon']))<img src="{{ $document['favicon'] }}" class="favicon" />@endif{{ $document['title'] }}</a>
+>@if(!empty($document['favicon']))<img src="{{ $document['favicon'] }}" class="favicon" />@endif {{ $document['title'] }}</a>
