@@ -1,3 +1,3 @@
 <a class="inline-document" 
     target="_blank" rel="noopener noreferrer" href="{{ $feed['url'] }}"
->@if(!empty($feed['favicon']))<img src="{{ $feed['favicon'] }}" class="favicon" />@endif{{ $feed['title'] }}</a>
+>@if(!empty($feed['favicon']))<img src="{{ $feed['favicon'] }}" class="favicon" />@endif {{ $feed['title'] }}</a>
