@@ -1,6 +1,8 @@
-<span class="inline-group">
-    <svg fill="currentColor" width="16" height="16" class="favicon folder-group inline">
-        <use xlink:href="{{ $iconsFileUrl }}#group" />
-    </svg>
-    {{ $group['name'] }}
-</span>
+<div class="inline-block rounded dark:bg-gray-800 px-2 py-1">
+    <div class="flex items-center space-x-1">
+        <svg fill="currentColor" width="16" height="16" class="text-green-500">
+            <use xlink:href="{{ asset('images/icons.svg') }}#group" />
+        </svg>
+        <span>{{ $group['name'] }}</span>
+    </div>
+</div>
